Fix user service - ensure action builder constructors are correct (were fetching users for all endpoints instead of contextual one) - removed duplicate cf-user-service
