+++ src/frontend/packages/store/src/effects/endpoint-api-errors.effects.ts
@@ -5,16 +5,10 @@ import { map } from 'rxjs/operators';
 
 import { SendEventAction } from '../actions/internal-events.actions';
 import { RequestTypes } from '../actions/request.actions';
+import { InternalAppState } from '../app-state';
 import { endpointSchemaKey } from '../helpers/entity-factory';
 import { InternalEventSeverity } from '../types/internal-events.types';
 import { WrapperRequestActionFailed } from '../types/request.types';
-import { InternalAppState } from '../app-state';
-
-
-
-// RequestTypes.FAILED
-
-
 
 @Injectable()
 export class EndpointApiError {
